Added a reducer test for state reset
